refactor: Remove unnecesery fields, interfaces, configs
