+++ package.json
@@ -18,8 +18,7 @@
     "dotenv": "^16.5.0",
     "express": "^4.21.2",
     "multer": "^1.4.5-lts.2",
-    "serverless-http": "^3.2.0",
-    "uuid": "^11.1.0"
+    "serverless-http": "^3.2.0"
   },
   "devDependencies": {
     "@types/aws-lambda": "^8.10.149",
