+++ src/config/multer-config.ts
@@ -6,15 +6,23 @@ const upload = multer({
   storage,
   limits: {
     fileSize: 5 * 1024 * 1024, // 5MB limit
+    files: 1, // Only allow single file upload
   },
-  /*fileFilter: (req, file, cb) => {
+  fileFilter: (req, file, cb) => {
     if (file.mimetype === 'text/csv' || file.originalname.endsWith('.csv')) {
       cb(null, true);
     } else {
       cb(new Error('Only CSV files are allowed'));
     }
   },
-  */
 });
 
-export const csvUpload = upload.array('csvFile');
+const multiUpload = multer({
+  storage,
+  limits: {
+    fileSize: 5 * 1024 * 1024, // 5MB limit
+  },
+});
+
+export const csvUpload = upload.single('csvFile');
+export const fileUploads = multiUpload.array('files');
