+++ src/modules/core/geonft.processor.ts
@@ -1,7 +1,5 @@
-import { FieldData } from '../../types';
-
 export class GeoNftProcessor {
-  validateAndTransform(): Promise<FieldData> {
-    return Promise.resolve({} as FieldData);
+  validateAndTransform(): Promise<unknown> {
+    return Promise.resolve({} as unknown);
   }
 }
