+++ src/modules/geonft/geonft.routes.ts
@@ -1,11 +1,11 @@
 import express from 'express';
 import { GeoNftController } from './geonft.controller';
-import { csvUpload } from '../../config/multer-config';
+import { fileUploads } from '../../config/multer-config';
 
 export function createGeoNftRoutes() {
   const router = express.Router();
   const controller = new GeoNftController();
 
-  router.post('/get-nft/process-csv', csvUpload, controller.handleGeoNftProcess);
+  router.post('/get-nft/process-csv', fileUploads, controller.handleGeoNftProcess);
   return router;
 }
